requirements.txt
Django==4.2.0
django-markdownx==4.0.2
Pillow==10.0.0
python-markdown==3.5.1